requirements.txt
asgiref
autopep8
django
pycodestyle
pytz
sqlparse
Unipath
dj-database-url
python-decouple
gunicorn
whitenoise
djangorestframework
channels
psycopg2==2.8.6
sentry-sdk
wagtail
django-bootstrap-datepicker-plus
django-tables2
django-filter
django-bootstrap3